Refactor UpdateAccountCard
